+++ src/index.ts
@@ -5,5 +5,5 @@ export * from './requestResponseTypes';
 export * from './dataTypes';
 export * from './helperTypes';
 export * from './plugins';
-export * from './SchedulerError';
+export * from './CronicleError';
 export * from './timingUtils';